Fix broken logic exposed by upstream fix
Fixed the method coding for the vertex exchange symmetry factor in
BMBPT diagrams. For some reason it was working but shouldn't by looking
at the NetworkX documentation, and something fixed that in NetworkX2.7.
The new version does not expect intesection to carry on the node info
(i.e. operator flag), as should be.